requirements.txt
fastapi==0.95.1
uvicorn==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
PyPDF2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==9.5.0
transformers==4.28.1
torch==2.0.0
faiss-cpu==1.7.4
numpy==1.24.3
scikit-learn==1.2.2
pytest==7.3.1
httpx==0.24.0
